requirements.txt
google-genai
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0